+++ lib/l10n/app_localizations_fr.dart
@@ -105,6 +105,9 @@ class AppLocalizationsFr extends AppLocalizations {
   String get repositories => 'Dépôts';
 
   @override
+  String get repositorySections => 'Repository Sections';
+
+  @override
   String get issues => 'Tickets';
 
   @override